refactor: renamed admin-page component to dashboard-page
