+++ src/app/views/dashboard-page/dashboard-page.component.html
@@ -1,4 +1,4 @@
-<div class="admin-page view-page grid spacing-20">
+<div class="dashboard-page view-page grid spacing-20">
   <div class="row">
     <div class="col">
       <h1>Admin settings</h1>
