+++ src/app/views/dashboard-page/dashboard-page.component.scss
@@ -1,7 +1,7 @@
 @use '../../../style/globals';
 @use '../../../style/colors';
 
-.admin-page {
+.dashboard-page {
   .display-icon .mat-icon {
     font-size: 48px;
     width: 48px;
