+++ src/app/views/dashboard-page/teams/teams-list-page/teams-list-page.component.scss
