+++ src/app/views/dashboard-page/teams/teams-view-page/teams-view-page.component.scss
