+++ src/app/views/dashboard-page/user-accounts-page/user-accounts-page.component.scss
